+++ App.js
@@ -1,7 +1,6 @@
 import 'react-native-gesture-handler';
 import React from "react";
 import MainStack from './Navigate';
-import { registerRootComponent } from 'expo';
 //import { StatusBar } from "react-native";
 //import { NativeBaseProvider } from 'native-base'
 import { AppContextProvider } from "./providers/AppContextProvider";
