+++ package.json
@@ -17,15 +17,14 @@
     "@react-navigation/native": "^6.1.7",
     "@react-navigation/native-stack": "^6.9.13",
     "@react-navigation/stack": "^6.3.17",
+    "@shopify/react-native-skia": "0.1.196",
     "expo": "^49.0.9",
     "expo-status-bar": "~1.6.0",
     "formik": "^2.4.2",
-    "lodash": "^4.17.21",
     "react": "^18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.72.5",
     "react-native-accordion-list-view": "^2.0.1",
-    "react-native-chart-kit": "^6.12.0",
     "react-native-charts-wrapper": "^0.5.11",
     "react-native-gesture-handler": "~2.12.0",
     "react-native-paper": "^5.8.0",
@@ -33,9 +32,7 @@
     "react-native-reanimated": "~3.3.0",
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0",
-    "react-native-web": "^0.19.6",
-    "victory": "^36.6.11",
-    "victory-native": "^40.0.2"
+    "react-native-web": "^0.19.6"
   },
   "devDependencies": {
     "@babel/core": "^7.22.5",
